optimize: move ModelAvailable
